Fikset nogle stier

+++ E1 1-8/src/Opg4/App.java
@@ -2,7 +2,7 @@ package Opg4;
 
 public class App {
     public static void main(String[] args) {
-        int balance = 1000;
+        double balance = 1000;
         for (int i = 0; i < 4; i++) {
             balance *= 1.05;
             System.out.println(balance);
